docs and cleanup exports
