+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "cm6-json-schema",
+  "name": "codemirror-json-schema",
   "license": "MIT",
   "version": "0.0.1",
   "description": "A JSONSchema enabled mode for codemirror 6, inspired by monaco-json",
@@ -17,9 +17,19 @@
   "packageManager": "pnpm@8.6.6",
   "types": "dist/index.d.ts",
   "module": "dist/index.js",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "types": "./dist/index.d.ts"
+    },
+    "./json5": {
+      "import": "./dist/json5.js",
+      "types": "./dist/json5.d.ts"
+    }
+  },
   "repository": {
     "type": "git",
-    "url": "https://github.com/acao/cm6-language-json-schema.git"
+    "url": "https://github.com/acao/codemirror-json-schema.git"
   },
   "dependencies": {
     "@changesets/cli": "^2.26.2",
