+++ src/debug.ts
@@ -1,5 +1,5 @@
 export const debug = {
   log: (...args: any[]) => {
     console.log(...args);
-  }
-}
+  },
+};
