+++ src/json-validation.ts
@@ -3,8 +3,9 @@ import type { Diagnostic } from "@codemirror/lint";
 import type { JSONSchema7 } from "json-schema";
 import { Draft04, type Draft, type JsonError } from "json-schema-library";
 import { joinWithOr } from "./utils/formatting";
-import { JSONPointerData } from "./utils/jsonPointers";
+import { JSONPointerData } from "./types";
 import { parseJSONDocumentState } from "./utils/parseJSONDocument";
+import { RequiredPick } from "./types";
 
 // return an object path that matches with the json-source-map pointer
 const getErrorPath = (error: JsonError): string => {
@@ -24,10 +25,12 @@ export type JSONValidationOptions = {
   formatError?: (error: JsonError) => string;
   jsonParser?: typeof parseJSONDocumentState;
 };
+
+type JSONValidationSettings = RequiredPick<JSONValidationOptions, "jsonParser">;
 /**
  * Helper for simpler class instantiaton
  */
-export function jsonSchemaLiner(
+export function jsonSchemaLinter(
   schema: JSONSchema7,
   options?: JSONValidationOptions
 ) {
@@ -39,7 +42,7 @@ export function jsonSchemaLiner(
 
 export class JSONValidation {
   private schema: Draft;
-  private options: JSONValidationOptions;
+  private options: JSONValidationSettings;
   public constructor(schema: JSONSchema7, options?: JSONValidationOptions) {
     this.options = {
       jsonParser: parseJSONDocumentState,
@@ -85,7 +88,7 @@ export class JSONValidation {
     // ignore blank json strings
     if (!text || text.trim().length < 3) return [];
 
-    const json = this.options.jsonParser!(view.state);
+    const json = this.options.jsonParser(view.state);
 
     let errors: JsonError[] = [];
     try {
