+++ src/json5.ts
@@ -1,0 +1,3 @@
+// json5
+export { json5SchemaLinter } from "./json5-validation";
+export { json5SchemaHover } from "./json5-hover";
